changes added by yatin

+++ src/main/java/com/ds/app/repository/iAppUserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.ds.app.entity.AppUser;
 
 @Repository
-public interface iAppUserRepository extends JpaRepository<AppUser,Integer>{
+public interface iAppUserRepository extends JpaRepository<AppUser,Long>{
 
 	public Optional<AppUser> findByUsername(String username);
